refactor: fix cypress methods for attribute group
