+++ cypress/defaults/index.js
@@ -2,15 +2,3 @@
  * Copyright © Ergonode Sp. z o.o. All rights reserved.
  * See LICENSE for license details.
  */
-export const DICTIONARIES = {
-    types: 'dictionary/attributes/types',
-    curr: 'dictionary/currencies',
-    units: 'units?view=list',
-    date: 'dictionary/date_format',
-    channels: 'dictionary/channels',
-    prodTypes: 'dictionary/product-type',
-    priv: 'dictionary/privileges',
-    sour: 'dictionary/sources',
-    lang: 'languages?limit=9999&offset=0&view=list&field=name&order=ASC',
-    langTree: 'language/tree',
-};
