+++ cypress/integration/common/navigation.test.js
@@ -12,107 +12,32 @@ import {
     MultiSteps,
 } from '../../models';
 import {
-    actionOnGrid,
     checkGridRow,
     noGridRow,
 } from '../../models/navigation';
-// import {
-//     getRequest,
-//     removeRequest,
-// } from '../../models/requests';
 
 MultiSteps([
-    When,
     Then,
     And,
-], 'I click on {string} button', (id) => {
+], 'Element {string} is visible', (id) => {
     cy.get(`[data-cy=${id}]`)
-        .click();
+        .should('be.visible');
 });
 
 MultiSteps([
-    When,
+    Then,
     And,
-], 'On {string} element I click button with {string} text', (id, text) => {
+], 'Element {string} is not visible', (id) => {
     cy.get(`[data-cy=${id}]`)
-        .find('button')
-        .contains(text)
-        .click();
+        .should('not.visible');
 });
 
 MultiSteps([
-    When,
     Then,
-], 'I fill the {string} input with the {string} term', (id, term) => {
+    And,
+], 'Element {string} not exist', (id) => {
     cy.get(`[data-cy=${id}]`)
-        .find('input')
-        .clear()
-        .type(term)
-        .should('have.value', term);
-});
-
-MultiSteps([
-    Then,
-    And,
-], 'I fill the {string} input with the {string} term for {string} translation', (id, term, language) => {
-    const name = `${id}_${language}`;
-
-    cy.get(`[data-cy=${name}]`)
-        .find('input')
-        .clear()
-        .type(term)
-        .should('have.value', term);
-});
-
-MultiSteps([
-    Then,
-    And,
-], 'I fill the {string} textarea with the {string} term for {string} translation', (id, term, language) => {
-    const name = `${id}_${language}`;
-
-    cy.get(`[data-cy=${name}]`)
-        .find('textarea')
-        .clear()
-        .type(term)
-        .should('have.value', term);
-});
-
-MultiSteps([
-    Then,
-    And,
-], 'I fill the {string} input for index {int} with the {string} term for {string} translation', (id, index, term, language) => {
-    const name = `${id}_${language}_${index}`;
-
-    cy.get(`[data-cy=${name}]`)
-        .find('input')
-        .clear()
-        .type(term)
-        .should('have.value', term);
-});
-
-MultiSteps([
-    Then,
-    When,
-], 'I click tab with {string} text', (text) => {
-    cy.get('[data-cy=tab-bar__items]')
-        .contains(text)
-        .click();
-});
-
-MultiSteps([
-    Then,
-    When,
-], 'I set {string} view on the grid', (view) => {
-    cy.get(`[data-cy=grid-${view}-view]`)
-        .click();
-});
-
-MultiSteps([
-    Then,
-    And,
-], 'Element {string} is {string}', (id, state) => {
-    cy.get(`[data-cy=${id}]`)
-        .should(`be.${state}`);
+        .should('not.exist');
 });
 
 MultiSteps([
@@ -120,37 +45,9 @@ MultiSteps([
     And,
 ], 'I can see {string} text on {string} element', (text, id) => {
     cy.get(`[data-cy=${id}]`)
-        .contains(text);
-});
-
-MultiSteps([
-    Then,
-    And,
-], 'I close modal', () => {
-    cy.get('[data-cy=modal]')
-        .should('be.visible');
-    cy.get('[data-cy=modal-close]')
-        .click();
-});
-
-MultiSteps([
-    Then,
-    And,
-], 'I close modal by button', () => {
-    cy.get('[data-cy=modal]')
-        .should('be.visible');
-    cy.get('[data-cy=modal-cancel]')
-        .click();
-});
-
-MultiSteps([
-    Then,
-    And,
-], 'I confirm modal', () => {
-    cy.get('[data-cy=modal]')
-        .should('be.visible');
-    cy.get('[data-cy=modal-confirm]')
-        .click();
+        .should(($ele) => {
+            expect($ele.text().trim()).equal(text);
+        });
 });
 
 MultiSteps([
@@ -164,82 +61,9 @@ MultiSteps([
 });
 
 MultiSteps([
-    Then,
-    And,
-], 'I click {string} check', (id) => {
-    cy.get(`[data-cy=${id}]`)
-        .find('label')
-        .click();
-});
-
-MultiSteps([
     And,
     Then,
-], 'I choose {int} option from {string} select field', (optionNr, id) => {
-    cy.get(`[data-cy=${id}]`)
-        .click();
-    cy.get(`[data-cy=${id}-drop-down]`)
-        .should('be.visible')
-        .find('.list > .list-element')
-        .as('elementList');
-    cy.get('@elementList')
-        .its('length')
-        .should('be.gt', 0);
-    cy.get('@elementList')
-        .eq(optionNr)
-        .as('selectedOption');
-    cy.get('@selectedOption')
-        .click({
-            force: true,
-        });
-    cy.get('@selectedOption')
-        .then(($option) => {
-            const optionValue = new RegExp($option.text()
-                .trim()
-                .replace('#', ''));
-
-            cy.get(`[data-cy=${id}-value] span`)
-                .contains(optionValue);
-        });
-    cy.get(`[data-cy=${id}-drop-down]`)
-        .should('be.not.visible');
-});
-
-MultiSteps([
-    And,
-    Then,
-], 'I choose {string} option(s) from {string} multi select field', (optionNrs, id) => {
-    const parsedOptions = JSON.parse(optionNrs);
-
-    cy.get(`[data-cy=${id}]`)
-        .click()
-        .should('be.visible');
-    cy.get(`[data-cy=${id}-drop-down]`)
-        .should('be.visible')
-        .find('.list > .list-element')
-        .as('elementList');
-    cy.get('@elementList')
-        .its('length')
-        .should('be.gt', 0);
-    cy.wrap(parsedOptions)
-        .each((optionNr) => {
-            cy.get('@elementList')
-                .eq(optionNr)
-                .as('selectedOption');
-            cy.get('@selectedOption')
-                .click();
-        });
-    cy.get(`[data-cy=${id}-drop-down]`)
-        .find('button')
-        .contains('OK')
-        .click();
-    cy.get(`[data-cy=${id}-drop-down]`)
-        .should('be.not.visible');
-});
-
-MultiSteps([
-    And,
-    Then,
+    When,
 ], 'On {string} I can see row with {string} value and columns data: {string}', (gridId, searchValue, columns) => {
     checkGridRow({
         gridId,
@@ -259,12 +83,23 @@ MultiSteps([
 });
 
 MultiSteps([
-    And,
     Then,
-], 'On {string} I click on {string} button for row with {string} value', (gridId, action, searchValue) => {
-    actionOnGrid({
-        gridId,
-        action,
-        searchValue,
-    });
+    And,
+], 'I see a form validation error that says {string}', (errors) => {
+    const parsedErrors = JSON.parse(errors.replace(/'/g, '"'));
+
+    cy
+        .get('[data-cy=form-errors]')
+        .should('be.visible')
+        .as('formErrorsWrapper');
+    cy
+        .get('@formErrorsWrapper')
+        .find('.errors-list__links > button')
+        .as('errorList');
+
+    cy.get('@errorList')
+        .should('have.length', parsedErrors.length)
+        .each(($error, i) => {
+            expect($error.find('span').text().trim()).to.equal(parsedErrors[i]);
+        });
 });
