+++ cypress/integration/errors/page_not_found.test.js
@@ -2,20 +2,20 @@
  * Copyright © Ergonode Sp. z o.o. All rights reserved.
  * See LICENSE for license details.
  */
-context('Page not found', () => {
-    before(() => {
-        cy.login(Cypress.env('adminEmail'), Cypress.env('adminPass'));
-        cy.visit('/users');
-    });
+// context('Page not found', () => {
+//     before(() => {
+//         cy.login(Cypress.env('adminEmail'), Cypress.env('adminPass'));
+//         cy.visit('/users');
+//     });
 
-    describe('Enter to invlide page URL', () => {
-        it('Invalid URL', () => {
-            cy.visit('page/not/found', {
-                failOnStatusCode: false,
-            });
-            cy.get('h1')
-                .invoke('text')
-                .should('include', 'Not found');
-        });
-    });
-});
+//     describe('Enter to invlide page URL', () => {
+//         it('Invalid URL', () => {
+//             cy.visit('page/not/found', {
+//                 failOnStatusCode: false,
+//             });
+//             cy.get('h1')
+//                 .invoke('text')
+//                 .should('include', 'Not found');
+//         });
+//     });
+// });
