+++ cypress/support/index.js
@@ -22,3 +22,7 @@ import './commands';
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
+
+Cypress.Cookies.defaults({
+    preserve: 'token',
+});
