+++ package.json
@@ -84,8 +84,8 @@
     "babel-loader": "^8.2.2",
     "babel-plugin-dynamic-import-node": "^2.3.3",
     "cross-env": "^7.0.3",
-    "cypress": "^4.12.1",
-    "cypress-cucumber-preprocessor": "^2.5.3",
+    "cypress": "^7.1.0",
+    "cypress-cucumber-preprocessor": "^4.0.3",
     "eslint": "^7.22.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-loader": "^4.0.2",
